// sensorregistration/graphql/gateway/src/grpc/generated/sensor.ts
export interface RegisterSensorRequest {
  sensorId: string;
  userEmail: string;
  postcode: string;
}

export interface GetSensorRequest {
  sensorId: string;
}

export interface ListSensorsByUserRequest {
  userEmail: string;
}

export interface UpdatePostcodeRequest {
  sensorId: string;
  newPostcode: string;
}

export interface SensorResponse {
  sensorId: string;
  userEmail: string;
  postcode: string;
  status: string;
  registeredAt: number;
  lastUpdatedAt: number;
}

export interface ListSensorsResponse {
  sensors: SensorResponse[];
}

function createBaseRegisterSensorRequest(): RegisterSensorRequest {
  return { sensorId: "", userEmail: "", postcode: "" };
}

export const RegisterSensorRequest = {
  encode(message: RegisterSensorRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.sensorId !== "") {
      writer.uint32(10).string(message.sensorId);
    }
    if (message.userEmail !== "") {
      writer.uint32(18).string(message.userEmail);
    }
    if (message.postcode !== "") {
      writer.uint32(26).string(message.postcode);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): RegisterSensorRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRegisterSensorRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.sensorId = reader.string();
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.userEmail = reader.string();
          continue;
        case 3:
          if (tag !== 26) {
            break;
          }

          message.postcode = reader.string();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): RegisterSensorRequest {
    return {
      sensorId: isSet(object.sensorId) ? globalThis.String(object.sensorId) : "",
      userEmail: isSet(object.userEmail) ? globalThis.String(object.userEmail) : "",
      postcode: isSet(object.postcode) ? globalThis.String(object.postcode) : "",
    };
  },

  toJSON(message: RegisterSensorRequest): unknown {
    const obj: any = {};
    if (message.sensorId !== "") {
      obj.sensorId = message.sensorId;
    }
    if (message.userEmail !== "") {
      obj.userEmail = message.userEmail;
    }
    if (message.postcode !== "") {
      obj.postcode = message.postcode;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<RegisterSensorRequest>, I>>(base?: I): RegisterSensorRequest {
    return RegisterSensorRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<RegisterSensorRequest>, I>>(object: I): RegisterSensorRequest {
    const message = createBaseRegisterSensorRequest();
    message.sensorId = object.sensorId ?? "";
    message.userEmail = object.userEmail ?? "";
    message.postcode = object.postcode ?? "";
    return message;
  },
};

function createBaseGetSensorRequest(): GetSensorRequest {
  return { sensorId: "" };
}

export const GetSensorRequest = {
  encode(message: GetSensorRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.sensorId !== "") {
      writer.uint32(10).string(message.sensorId);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): GetSensorRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetSensorRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.sensorId = reader.string();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetSensorRequest {
    return { sensorId: isSet(object.sensorId) ? globalThis.String(object.sensorId) : "" };
  },

  toJSON(message: GetSensorRequest): unknown {
    const obj: any = {};
    if (message.sensorId !== "") {
      obj.sensorId = message.sensorId;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GetSensorRequest>, I>>(base?: I): GetSensorRequest {
    return GetSensorRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GetSensorRequest>, I>>(object: I): GetSensorRequest {
    const message = createBaseGetSensorRequest();
    message.sensorId = object.sensorId ?? "";
    return message;
  },
};

function createBaseListSensorsByUserRequest(): ListSensorsByUserRequest {
  return { userEmail: "" };
}

export const ListSensorsByUserRequest = {
  encode(message: ListSensorsByUserRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.userEmail !== "") {
      writer.uint32(10).string(message.userEmail);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): ListSensorsByUserRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseListSensorsByUserRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.userEmail = reader.string();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ListSensorsByUserRequest {
    return { userEmail: isSet(object.userEmail) ? globalThis.String(object.userEmail) : "" };
  },

  toJSON(message: ListSensorsByUserRequest): unknown {
    const obj: any = {};
    if (message.userEmail !== "") {
      obj.userEmail = message.userEmail;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ListSensorsByUserRequest>, I>>(base?: I): ListSensorsByUserRequest {
    return ListSensorsByUserRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ListSensorsByUserRequest>, I>>(object: I): ListSensorsByUserRequest {
    const message = createBaseListSensorsByUserRequest();
    message.userEmail = object.userEmail ?? "";
    return message;
  },
};

function createBaseUpdatePostcodeRequest(): UpdatePostcodeRequest {
  return { sensorId: "", newPostcode: "" };
}

export const UpdatePostcodeRequest = {
  encode(message: UpdatePostcodeRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.sensorId !== "") {
      writer.uint32(10).string(message.sensorId);
    }
    if (message.newPostcode !== "") {
      writer.uint32(18).string(message.newPostcode);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): UpdatePostcodeRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseUpdatePostcodeRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.sensorId = reader.string();
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.newPostcode = reader.string();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): UpdatePostcodeRequest {
    return {
      sensorId: isSet(object.sensorId) ? globalThis.String(object.sensorId) : "",
      newPostcode: isSet(object.newPostcode) ? globalThis.String(object.newPostcode) : "",
    };
  },

  toJSON(message: UpdatePostcodeRequest): unknown {
    const obj: any = {};
    if (message.sensorId !== "") {
      obj.sensorId = message.sensorId;
    }
    if (message.newPostcode !== "") {
      obj.newPostcode = message.newPostcode;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<UpdatePostcodeRequest>, I>>(base?: I): UpdatePostcodeRequest {
    return UpdatePostcodeRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<UpdatePostcodeRequest>, I>>(object: I): UpdatePostcodeRequest {
    const message = createBaseUpdatePostcodeRequest();
    message.sensorId = object.sensorId ?? "";
    message.newPostcode = object.newPostcode ?? "";
    return message;
  },
};
// ...
export const ListSensorsResponse = {
  encode(message: ListSensorsResponse, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    for (const v of message.sensors) {
      SensorResponse.encode(v!, writer.uint32(10).fork()).ldelim();
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): ListSensorsResponse {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseListSensorsResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.sensors.push(SensorResponse.decode(reader, reader.uint32()));
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ListSensorsResponse {
    return {
      sensors: globalThis.Array.isArray(object?.sensors)
        ? object.sensors.map((e: any) => SensorResponse.fromJSON(e))
        : [],
    };
  },

  toJSON(message: ListSensorsResponse): unknown {
    const obj: any = {};
    if (message.sensors?.length) {
      obj.sensors = message.sensors.map((e) => SensorResponse.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ListSensorsResponse>, I>>(base?: I): ListSensorsResponse {
    return ListSensorsResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ListSensorsResponse>, I>>(object: I): ListSensorsResponse {
    const message = createBaseListSensorsResponse();
    message.sensors = object.sensors?.map((e) => SensorResponse.fromPartial(e)) || [];
    return message;
  },
};

export interface SensorService {
  RegisterSensor(request: RegisterSensorRequest): Promise<SensorResponse>;
  GetSensor(request: GetSensorRequest): Promise<SensorResponse>;
  ListSensorsByUser(request: ListSensorsByUserRequest): Promise<ListSensorsResponse>;
  UpdateSensorPostcode(request: UpdatePostcodeRequest): Promise<SensorResponse>;
}

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };